Move Biscuit Lifecycle files to a separate folder
