+++ src/components/Machine/index.js
@@ -2,7 +2,7 @@ import { useEffect, useState } from "react";
 import { Container, Row, Col } from "react-bootstrap";
 
 import * as constants from "../../constants.js";
-import BiscuitLifecycle from "../Biscuit/biscuitLifecycle.js";
+import BiscuitLifecycle from "../Biscuit/BiscuitLifecycle";
 import Stamper from "../Stamper";
 import Switch from "../Switch";
 import BoostSwitch from "../boostSwitch.js";
